UI State Management Update
These are the controllers that enable the state transition between the different views. Edit hellocontroller to change where you are inside the user dashboard or the admin dashboard. Got nav menus working as well, including a functional "quit" menu option. Will be cleaning this up and making it more consistent. Feel free to give it a whirl