Removed menu boilerplate, added about app/metadata
When following the JavaFX tutorial, the template I used included a lot of menu item boilerplate that we are not using. New FXML files reflect current menu items in use with IDs. Also fixed a few labels as I revisited certain views. Also added a new view that is intended as the hub for support documentation.