Added create/delete/edit functionality & other fixes
Added create, delete, and edit functionality for items and collections. Removed many TODO comments and unused methods. Includes new logic for added search buttons I forgot to add to some views as well. Also includes an initialize method for the login scene to check if users are in the DB, and if not enable FTS button to prevent softlock.